+++ src/main/java/edu/ie3/datamodel/io/source/EntitySource.java
@@ -204,14 +204,12 @@ public abstract class EntitySource {
    * @return a stream of the entity data wrapped in a {@link Try}
    */
   protected static Stream<Try<EntityData, SourceException>> buildEntityData(
-      Class<? extends Entity> entityClass, DataSource dataSource) {
-    return Try.of(() -> dataSource.getSourceData(entityClass), SourceException.class)
-        .convert(
-            data ->
-                data.map(
-                    fieldsToAttributes ->
-                        new Try.Success<>(new EntityData(fieldsToAttributes, entityClass))),
-            exception -> Stream.of(Failure.of(exception)));
+      Class<? extends Entity> entityClass, DataSource dataSource) throws SourceException {
+    return dataSource
+        .getSourceData(entityClass)
+        .map(
+            fieldsToAttributes ->
+                new Try.Success<>(new EntityData(fieldsToAttributes, entityClass)));
   }
 
   /**
@@ -227,7 +225,8 @@ public abstract class EntitySource {
   protected static <E extends EntityData> Stream<Try<E, SourceException>> buildEntityData(
       Class<? extends Entity> entityClass,
       DataSource dataSource,
-      WrappedFunction<EntityData, E> converter) {
+      WrappedFunction<EntityData, E> converter)
+      throws SourceException {
     return buildEntityData(entityClass, dataSource).map(converter);
   }
 
@@ -356,9 +355,7 @@ public abstract class EntitySource {
    */
   protected static <S, E extends Exception> Stream<S> unpack(
       Stream<Try<S, E>> inputStream, Class<S> clazz) throws SourceException {
-    return Try.scanStream(inputStream, clazz.getSimpleName())
-        .transformF(SourceException::new)
-        .getOrThrow();
+    return Try.scanStream(inputStream, clazz.getSimpleName(), SourceException::new).getOrThrow();
   }
 
   /**
@@ -376,16 +373,14 @@ public abstract class EntitySource {
     return entityData.flatMap(
         data ->
             Try.of(() -> data.getUUID(fieldName), FactoryException.class)
+                .flatMap(entityUuid -> extractFunction(entityUuid, entities))
                 .transformF(
                     exception ->
                         new SourceException(
-                            "Extracting UUID field "
+                            "Extracting UUID for field '"
                                 + fieldName
-                                + " from entity data "
-                                + entityData
-                                + " failed.",
-                            exception))
-                .flatMap(entityUuid -> extractFunction(entityUuid, entities)));
+                                + "' failed. Caused by: "
+                                + exception.getMessage())));
   }
 
   /**
@@ -396,13 +391,13 @@ public abstract class EntitySource {
    * @return a try of the {@link Entity}
    * @param <T> type of entity
    */
-  protected static <T> Try<T, SourceException> extractFunction(UUID uuid, Map<UUID, T> entityMap) {
+  protected static <T> Try<T, FactoryException> extractFunction(UUID uuid, Map<UUID, T> entityMap) {
     return Optional.ofNullable(entityMap.get(uuid))
         // We either find a matching entity for given UUID, thus return a success
-        .map(entity -> Try.of(() -> entity, SourceException.class))
+        .map(entity -> Try.of(() -> entity, FactoryException.class))
         // ... or find no matching entity, returning a failure.
         .orElse(
-            new Failure<>(new SourceException("Entity with uuid " + uuid + " was not provided.")));
+            new Failure<>(new FactoryException("Entity with uuid " + uuid + " was not provided.")));
   }
 
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=